removed tdfs and updated docs
