+++ featuretools/__init__.py
@@ -10,6 +10,7 @@ from .primitives import Feature
 from .computational_backends.api import *
 from . import tests
 from .utils.pickle_utils import *
+from .utils.time_utils import *
 import featuretools.demo
 
 __version__ = '0.1.19'
