+++ featuretools/synthesis/api.py
@@ -3,5 +3,4 @@ from __future__ import absolute_import
 # flake8: noqa
 from .deep_feature_synthesis import DeepFeatureSynthesis
 from .dfs import dfs
-from .tdfs import tdfs, make_temporal_cutoffs
 from .encode_features import encode_features
