+++ featuretools/utils/api.py
@@ -1,4 +1,5 @@
 # flake8: noqa
 from .gen_utils import make_tqdm_iterator
 from .pickle_utils import load_features, save_features
+from .time_utils import make_temporal_cutoffs
 from .wrangle import flatten_2d
